Update glMapBuffer.xhtml
GL_BUFFER_USAGE is not a constant or enum. The original "BUFFER_USAGE" name is part of the internal state of a buffer object, and comes from the original text found in the Opengl spec document. See Table 6.3 in the Opengl v4.6 spec.
If this update is accepted, it should be done on glMapBufferRange as well.