Fix compile process for python3 on Windows
As it stands python3 will fail to read in the input xhtml files
because they contain bytes that are not valid using the default
codepage on Windows (in English at least). Loading those files
as utf-8 explicitly resolves this issue and is correct as long
as all the xhtml files are utf-8 encoded.